+++ evap/evaluation/models.py
@@ -38,7 +38,6 @@ from evap.evaluation.tools import (
     translate,
     vote_end_datetime,
 )
-from evap.settings import PAGE_URL
 
 logger = logging.getLogger(__name__)
 
@@ -958,23 +957,6 @@ class Evaluation(LoggedModel):
             num_participants=Subquery(participant_count_subquery),
             num_voters=Subquery(voter_count_subquery),
         )
-
-    @classmethod
-    def get_sorted_evaluation_url_tuples_with_urgent_review(cls) -> List[Tuple["Evaluation", str]]:
-        evaluation_url_tuples = [
-            (
-                evaluation,
-                PAGE_URL
-                + reverse(
-                    "staff:evaluation_textanswers",
-                    kwargs={"semester_id": evaluation.course.semester.id, "evaluation_id": evaluation.id},
-                ),
-            )
-            for evaluation in Evaluation.objects.filter(state=Evaluation.State.EVALUATED)
-            if evaluation.textanswer_review_state == Evaluation.TextAnswerReviewState.REVIEW_URGENT
-        ]
-        evaluation_url_tuples = sorted(evaluation_url_tuples, key=lambda evaluation: evaluation[0].full_name)
-        return evaluation_url_tuples
 
     @property
     def unlogged_fields(self):
